requirements.txt
nba_api
pandas
sentence-transformers
chromadb
fastapi
uvicorn
python-dotenv
google-generativeai